Refactor: service 폴더구조 변경 및 interface 생성
