+++ src/main/java/com/example/travelhana/Config/SecurityConfig.java
@@ -50,7 +50,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 				"/swagger-resources/**",
 				"/v2/api-docs",
 				"/webjars/**",
-				"/signup/**", "/signin/password/**", "/refresh/**", "/userrole/**", "/verification/auth", "/verification",
+				"/signup/**", "/signin/password/**","/registration/**", "/refresh/**", "/userrole/**", "/verification/auth", "/verification",
                 "/account/dummy", "marker/dummy").permitAll();
 
 		http.authorizeRequests().anyRequest().authenticated();
