+++ src/main/java/com/example/travelhana/Controller/ExchangeController.java
@@ -2,9 +2,7 @@ package com.example.travelhana.Controller;
 
 import com.example.travelhana.Dto.ExchangeRequestDto;
 import com.example.travelhana.Service.ExchangeService;
-import io.swagger.models.Response;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
 import org.springframework.web.bind.annotation.PostMapping;
