+++ src/main/java/com/example/travelhana/Service/implement/CustomUserDetailsImpl.java
@@ -1,15 +1,9 @@
-package com.example.travelhana.Service.Implement;
+package com.example.travelhana.Service.implement;
 
 import com.example.travelhana.Service.CustomUserDetails;
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
 
 import java.util.Collection;
-import java.util.List;
 
 public class CustomUserDetailsImpl implements CustomUserDetails {
